refactor: apply instructor feedback

+++ src/main/java/com/theo/community_api/post/service/PostService.java
@@ -2,10 +2,10 @@ package com.theo.community_api.post.service;
 
 import com.theo.community_api.comment.domain.Comment;
 import com.theo.community_api.comment.repository.CommentRepository;
-import com.theo.community_api.common.exception.ForbiddenException;
-import com.theo.community_api.common.exception.NotFoundException;
+import com.theo.community_api.common.exception.*;
 import com.theo.community_api.post.domain.Post;
 import com.theo.community_api.post.dto.*;
+import com.theo.community_api.post.repository.PostReportRepository;
 import com.theo.community_api.post.repository.PostRepository;
 import com.theo.community_api.reply.domain.Reply;
 import com.theo.community_api.reply.repository.ReplyRepository;
@@ -25,6 +25,7 @@ public class PostService {
     private final UserRepository userRepository;
     private final CommentRepository commentRepository;
     private final ReplyRepository replyRepository;
+    private final PostReportRepository postReportRepository;
 
     // 게시글 추가
     public Long createPost(Long loginUserId, PostCreateRequest request) {
@@ -53,15 +54,18 @@ public class PostService {
         List<PostSummaryResponse> posts = new ArrayList<>();
 
         for (Post post : allPosts) {
-            // 추가 조회 시 lastPostId보다 작은 게시글만 조회
+            if (post.isDeleted()) { // soft delete 된 게시물은 안보이도록 하기
+                continue;
+            }
+            // 추가 조회시 lastPostId보다 작은 게시물부터 조회
             if (lastPostId != null && post.getPostId() >= lastPostId) {
                 continue;
             }
+
             User user = userRepository.findById(post.getUserId()).orElse(null);
 
             posts.add(PostSummaryResponse.from(post, user));
 
-            // hasNext 확인을 위해 size + 1개까지 조회
             if (posts.size() == size + 1) {
                 break;
             }
@@ -89,10 +93,14 @@ public class PostService {
     // 게시글 상세 조회
     public PostDetailResponse readPostDetail(Long postId) {
         Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new NotFoundException("post_not_found"));
+                .orElseThrow(() -> new BusinessException(ErrorCode.POST_NOT_FOUND));
+
+        if (post.isDeleted()) { // soft delete 된 게시물 안보이도록 설정
+            throw new BusinessException(ErrorCode.POST_NOT_FOUND);
+        }
 
         if (post.isBlinded()) {
-            throw new ForbiddenException("post_blinded");
+            throw new BusinessException(ErrorCode.POST_BLINDED);
         }
 
         post.increaseViewCount();
@@ -104,13 +112,28 @@ public class PostService {
         List<PostCommentResponse> commentResponses = new ArrayList<>();
 
         for (Comment comment : comments) { // 댓글 리스트 가져오기
+            if (comment.isDeletedByPost()) {
+                continue;
+            }
+
             User commentAuthor = userRepository.findById(comment.getUserId())
                     .orElse(null);
 
             List<Reply> replies = replyRepository.findAllByCommentId(comment.getCommentId());
+
             List<PostReplyResponse> replyResponses = new ArrayList<>();
 
             for (Reply reply : replies) { // 대댓글 리스트 가져오기
+                if (reply.isDeleted()) {
+                    continue;
+                }
+
+                // 댓글이 삭제됐고, 살아있는 대댓글도 없으면 화면에서 숨긴다
+                if (comment.isCommentDeleted() && replyResponses.isEmpty()) {
+                    continue;
+                }
+
+
                 User replyAuthor = userRepository.findById(reply.getUserId())
                         .orElse(null);
 
@@ -130,9 +153,9 @@ public class PostService {
     // 게시글 수정
     public PostUpdateResponse updatePost(Long loginUserId, Long postId, PostUpdateRequest request) {
         Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new NotFoundException("post_not_found"));
+                .orElseThrow(() -> new BusinessException(ErrorCode.POST_NOT_FOUND));
         if (!post.getUserId().equals(loginUserId)) {
-            throw new ForbiddenException("post_modify_forbidden");
+            throw new BusinessException(ErrorCode.POST_MODIFY_FORBIDDEN);
         }
         post.update(request.getTitle(), request.getContent(), request.getPostImage());
 
@@ -142,19 +165,51 @@ public class PostService {
     // 게시글 삭제
     public void deletePost(Long loginUserId, Long postId) {
         Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new NotFoundException("post_not_found"));
+                .orElseThrow(() -> new BusinessException(ErrorCode.POST_NOT_FOUND));
+
+        if (post.isDeleted()){
+            throw new BusinessException(ErrorCode.POST_NOT_FOUND);
+        }
 
         if (!post.getUserId().equals(loginUserId)) {
-            throw new ForbiddenException("post_delete_forbidden");
+            throw new BusinessException(ErrorCode.POST_DELETE_FORBIDDEN);
         }
 
         postRepository.deleteById(postId);
+
+        List<Comment> comments = commentRepository.findAllByPostId(postId);
+
+        for (Comment comment : comments) {
+            comment.deleteByPostDeleted();
+
+            List<Reply> replies =
+                    replyRepository.findAllByCommentId(comment.getCommentId());
+
+            for (Reply reply : replies) {
+                reply.delete();
+            }
+        }
     }
 
     // 게시글 신고
     public void reportPost(Long loginUserId, Long postId) {
         Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new NotFoundException("post_not_found"));
+                .orElseThrow(() -> new BusinessException(ErrorCode.POST_NOT_FOUND));
+
+        if(post.isDeleted()){
+            throw new BusinessException(ErrorCode.POST_NOT_FOUND);
+        }
+
+        if(post.isBlinded()){
+            throw new BusinessException(ErrorCode.POST_BLINDED);
+        }
+
+        // 한 사용자가 같은 글에 중복되서 신고한 경우
+        if(postReportRepository.existsByPostIdAndUserId(postId, loginUserId)){
+            throw new BusinessException(ErrorCode.DUPLICATE_REPORT);
+        }
+
+        postReportRepository.reportSave(postId, loginUserId);
 
         post.increaseReportCount();
     }
